+++ src/Facades/Finam.php
@@ -6,9 +6,11 @@ namespace MasyaSmv\FinamSdk\Facades;
 
 use Illuminate\Support\Facades\Facade;
 use MasyaSmv\FinamSdk\Client\FinamClient;
+use MasyaSmv\FinamSdk\Dto\Auth\IssuedTokenDto;
 use MasyaSmv\FinamSdk\Contracts\FinamSessionInterface;
 
 /**
+ * @method static IssuedTokenDto issueToken(string $secret)
  * @method static FinamSessionInterface connect(string $token)
  * @method static FinamClient client(?string $token = null)
  *